Model inheritance optimisation.
Always select the `content_type` related field
when selecting subclasses since it will be
needed to type cast correctly.

Note that the `FieldDefinition.field_type` field
was renamed to `content_type` in order to maintain
good abstraction in `InheritanceManager`. I'm
thinking of moving this whole layer to a separated
app since it's getting pretty mature.