Fixed #19 -- Avoid forcing model creation when not required.
Since the model class is solely used to determine the table name
and the migrateable database it's safe to assume it doesn't need
to be recreated when loading a raw field definitions.

The forced model creation caused an issue with non-proxy field definition
subclasses on fixture loading because their associated subclass details
we're not yet created and thus couldn't be type casted by polymodels.

The test fixture order was changed in order to reproduce the scenario
of a proxy field definition instance being loaded before the concrete
one data associated with a previously loaded field definition.

Thanks to @caneta for the detailed report.